requirements.txt
langchain-chroma
flask==3.0.*
gunicorn==22.0.*
gevent==24.*
pillow
pytesseract
pymupdf
tiktoken
google-generativeai
langchain
langchain-community
chromadb
faiss-cpu
sentence-transformers
docling
tqdm
